allowing binary parameters to be IOSchemas
